# services/assessment-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_jjaenal_sisfo-akademik-backend_services_assessment-service_internal_domain_entity.ReportCardTemplate:
    properties:
      config:
        $ref: '#/definitions/github_com_jjaenal_sisfo-akademik-backend_services_assessment-service_internal_domain_entity.TemplateConfig'
      created_at:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  github_com_jjaenal_sisfo-akademik-backend_services_assessment-service_internal_domain_entity.TemplateConfig:
    properties:
      footer_text:
        type: string
      header_text:
        type: string
      logo_url:
        type: string
      primary_color:
        description: Hex code
        type: string
      secondary_color:
        description: Hex code
        type: string
    type: object
  httputil.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  httputil.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/httputil.ErrorDetail'
      meta:
        $ref: '#/definitions/httputil.Meta'
      success:
        example: false
        type: boolean
    type: object
  httputil.Meta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
host: localhost:8084
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Assessment Service for Sisfo Akademik
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Assessment Service API
  version: "1.0"
paths:
...
  /report-cards/{id}/download:
    get:
      description: Download the generated report card PDF
      parameters:
      - description: Report Card ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Download report card PDF
      tags:
      - report-cards
  /report-cards/generate:
    post:
      consumes:
      - application/json
      description: Generate a report card for a student
      parameters:
      - description: Generate Request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Generate a report card
      tags:
      - report-cards
  /report-cards/student/{student_id}:
    get:
      consumes:
      - application/json
      description: Get report card for a student in a semester
      parameters:
      - description: Student ID
        in: path
        name: student_id
        required: true
        type: string
      - description: Semester ID
        in: query
        name: semester_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Get report card by student
      tags:
      - report-cards
  /templates:
    get:
      consumes:
      - application/json
      description: List report card templates by tenant ID
      parameters:
      - description: Tenant ID
        in: query
        name: tenant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_jjaenal_sisfo-akademik-backend_services_assessment-service_internal_domain_entity.ReportCardTemplate'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: List templates
      tags:
      - templates
    post:
      consumes:
      - application/json
      description: Create a new report card template
      parameters:
      - description: Template Request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jjaenal_sisfo-akademik-backend_services_assessment-service_internal_domain_entity.ReportCardTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Create a template
      tags:
      - templates
  /templates/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a report card template
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Delete a template
      tags:
      - templates
    get:
      consumes:
      - application/json
      description: Get a report card template by ID
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jjaenal_sisfo-akademik-backend_services_assessment-service_internal_domain_entity.ReportCardTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Get a template
      tags:
      - templates
    put:
      consumes:
      - application/json
      description: Update a report card template
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jjaenal_sisfo-akademik-backend_services_assessment-service_internal_domain_entity.ReportCardTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Update a template
      tags:
      - templates
schemes:
- http
swagger: "2.0"